// types/formation-id.ts
export type FormationId =
    | "3412"
    | "3421"
    | "3142"
    | "343"
    | "352"
    | "41212"
    | "41212_2"
    | "4141"
    | "4231"
    | "4231_2"
    | "4222"
    | "424"
    | "4312"
    | "4132"
    | "4321"
    | "433"
    | "433_2"
    | "433_3"
    | "433_4"
    | "433_5"
    | "4411"
    | "4411_2"
    | "442"
    | "442_2"
    | "451"
    | "451_2"
    | "5212"
    | "5221"
    | "532"
    | "541";